Rename database to db
Resolves #99
