+++ server/auth/index.js
@@ -2,7 +2,7 @@ import express from 'express';
 import config from '../config/environment';
 import facebookSetup from './facebook/passport';
 import facebookRouter from './facebook';
-import { User } from '../database/database';
+import { User } from '../db';
 
 facebookSetup(User, config);
